requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.11.1
    # via django
build==1.4.3
    # via pip-tools
click==8.3.2
    # via pip-tools
django==6.0.4
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.17.1
    # via -r requirements.in
packaging==26.1
    # via
    #   build
    #   wheel
pip-tools==7.5.3
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
sqlparse==0.5.5
    # via django
wheel==0.46.3
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
